Refactor current shared exclusion for Vac Sources
- Move the shared code from model concerns to the VacancySource namespace.
- Move constant definition to School model (will be needed outside this
  namespace).
- Try to reuse existing function interface instead of overwriting it for
  individual school.
