+++ app/models/school.rb
@@ -6,6 +6,21 @@ class School < Organisation
 
   validates :urn, uniqueness: true
 
+  EXCLUDED_DETAILED_SCHOOL_TYPES = [
+    "Further education",
+    "Other independent school",
+    "Online provider",
+    "British schools overseas",
+    "Institution funded by other government department",
+    "Miscellaneous",
+    "Offshore schools",
+    "Service children’s education",
+    "Special post 16 institution",
+    "Other independent special school",
+    "Higher education institutions",
+    "Welsh establishment",
+  ].freeze
+
   READABLE_PHASE_MAPPINGS = {
     not_applicable: nil,
     nursery: "nursery",
