+++ app/vacancy_sources/vacancy_source/source/ark.rb
@@ -1,6 +1,6 @@
 class VacancySource::Source::Ark
   include VacancySource::Parser
-  include VacancySourceShared
+  include VacancySource::Shared
 
   FEED_URL = ENV.fetch("VACANCY_SOURCE_ARK_FEED_URL").freeze
   SOURCE_NAME = "ark".freeze
